GTK: use cairo_paint() to copy from backing store to screen
I just found cairo_paint() while wandering the Cairo documentation.  It
just fills the entire clip region with data from the current source,
which is precisely what draw_area() wants to do.  This is simpler for us
than requesting the bounding rectangle of the clipping region and then
filling it, and as far as I can tell the clipping rectangle generally
covers the whole window anyway.