GTK: disable GTK double-buffering of terminal drawing area
By default, when we're asked to draw a GTK widget, GTK creates a
temporary surface (a Pixmap under X) and redirects our rendering into
that.  Then it blits that into the actual window.  This is silly for
PuTTY because all that PuTTY does to render its drawing area is to
blit into it from _another_ surface.  So now PuTTY asks GTK not to do
that.  According to the GTK documentation, GTK as of 3.10 has
completely restructured its drawing routines so that turning off
double-buffering actually makes things worse and slower, so we turn it
off only in GTK 2

Still, this now means that painting text on the screen in GTK 2 causes
precisely one CopyArea operation, which is what we want.