GTK: correct a misuse of WhitePixel and BlackPixel
According to the X specs, WhitePixel and BlackPixel refer to permanent
entries in the default colourmap.  This means that they're not
necessarily appropriate for use with a Drawable with a different depth
than the root window.  When drawing to a Pixmap that will be used as a
1-bit alpha mask by Cairo, the correct values are simply 0
(transparent) and 1 (opaque).