GTK: purely server-side X bitmap font rendering with Cairo
This is a fairly radical change of how X bitmap fonts are handled when
using Cairo for rendering.  Before, we would download each glyph to the
client on first use and then composite those glyphs into the terminal's
backing surface. This worked pretty well when we were keeping an image
of the whole screen on the client anyway, but once I'd pushed all the
other Cairo rendering onto the X server, it meant that the character
bitmaps had to be repeatedly pushed to the X server.

The new arrangement just renders each string into a temporary Pixmap
using the usual X text-drawing calls and then asks Cairo to paste it
into the main backing Pixmap.  It's tempting to draw the text straight
into the backing Pixmap, but that would require dealing directly with
X colour management.  This way, we get to leave colours in the hands
of Cairo (and hence the Render extension).

There are still fragments of the old system around.  Those should go
in the next commit.