HTTP proxy: fix misspelled error message (and comment).
A 407 response from the proxy is expected to contain at least one
header naming an acceptable authentication method. The header name is
"Proxy-Authenticate". But if no such header is seen, our error message
incorrectly spelled the missing header name as "Proxy-Authorization".

That's a braino rather than a typo: "Proxy-Authorization" is the
header the _client_ sends for an actual attempt to authenticate.