HTTP proxy digest auth: tolerate multiple qop options.
We were rejecting any 'Proxy-Authenticate: Digest' header from the
proxy if the 'qop' setting was not exactly the string "auth", which is
the only quality of protection that this code supports. However 'qop'
is supposed to be a _list_ of possibilities, so we should also be
happy if the proxy presents a list of more than one option, as long as
"auth" is one of them.