FAQ: update website/hosting/forges section.
The question about a nicer domain name, with answer "we don't want
one", is very out of date now that we actually have one! Also, the
section about putty.org wants updating after the recent drama. While I
was there I decided the Sourceforge question was too old to live, and
replaced it with some more up-to-date thoughts about forges in general
and Github in particular.

The new question faq-putty-org-pivot ("putty.org became weird kookery,
have you been hacked or bought out?") is frequently asked in the sense
that I've seen it twice in the past week or two: once in email, and
once on Mastodon (though someone quickly posted the right answer in
response).

While I'm here, it seems a bit pointless these days for faq-link to
explain the Google pagerank system to everyone. It's been a long time
since Google Search was new enough that it needed explaining, and in
any case, I've no idea if it's still accurate any more!